config.py
import os
from dotenv import load_dotenv

load_dotenv()

# API Keys (store in .env file, never hardcode)
ANTHROPIC_API_KEY = os.getenv("ANTHROPIC_API_KEY")
NOTION_TOKEN      = os.getenv("NOTION_TOKEN")
NOTION_PAGE_ID    = os.getenv("NOTION_PAGE_ID")
SENDGRID_KEY      = os.getenv("SENDGRID_KEY")
SLACK_WEBHOOK     = os.getenv("SLACK_WEBHOOK_URL")
CRUNCHBASE_KEY    = os.getenv("CRUNCHBASE_API_KEY")

# ── RSS sources (free, no auth needed) ──────────────────────────
RSS_FEEDS = [
    # Global robotics
    "https://www.therobotreport.com/feed/",
    "https://roboticstomorrow.com/rss/articles",
    "https://spectrum.ieee.org/feeds/topic/robotics.rss",
    # Asia / SEA focused
    "https://www.futureiot.tech/feed/",
    "https://www.techinasia.com/feed",
    "https://e27.co/feed/",
    # Funding / startup
    "https://techcrunch.com/tag/robotics/feed/",
    "https://venturebeat.com/category/ai/feed/",
]

# ── Keyword filter — only keep articles matching these ──────────
KEYWORDS = [
    "robot", "robotics", "cobot", "collaborative robot", "AMR",
    "autonomous mobile robot", "humanoid", "automation",
    "industrial automation", "vietnam", "southeast asia", "ASEAN",
    "manufacturing", "warehouse", "logistics automation",
    "applied robotics", "computer vision", "machine vision",
]

# ── VinDynamic analyst persona (injected into every Claude call) ─
ANALYST_PERSONA = """
You are a senior market research analyst at VinDynamic, a Vietnamese 
applied robotics startup. Your job is to produce a concise, 
actionable daily intelligence briefing for the head of market research.

VinDynamic's focus areas:
- Applied robotics for manufacturing, logistics, and food processing
- Primary markets: Vietnam, Southeast Asia; secondary: global
- Key competitors to watch: ABB, KUKA, Doosan Robotics, Techman Robot,
  Universal Robots, local Vietnamese integrators
- Priority verticals: electronics manufacturing, seafood/food processing,
  warehouse/logistics, construction inspection

Your briefing style:
- Lead with the most strategically important insight
- Flag anything relevant to Vietnam or SEA specifically
- Note competitor moves and funding events
- Use plain business language, no jargon
- Be specific: name companies, cite numbers, state implications
"""

# ── Format báo cáo bắt buộc (REPORT FORMAT) ─────────────────────
REPORT_FORMAT = """
Produce a daily intelligence briefing with EXACTLY these sections:

## Market trends
2-4 bullet points on broader robotics market movements.

## Competitor & industry moves
2-4 bullet points on specific company actions, product launches, partnerships.

## Funding & M&A
1-3 bullet points on investment rounds, acquisitions, exits.
(Write "Nothing significant today." if none found.)

## Southeast Asia & Vietnam spotlight
2-3 bullet points specifically relevant to SEA or Vietnam.
(Write "No SEA-specific news today." if none found.)

## Strategic implications for VinDynamic
3 bullet points: what does today's news mean for VinDynamic specifically?
Be direct and actionable.

## Top 3 articles to read in full
List the 3 most important articles with their URLs.

End with a one-sentence "headline of the day" that captures the single
most important development.
"""
